p2p: fix some stuff after refactor
License: MIT
