+++ p2p/p2p.go
@@ -16,6 +16,22 @@ type P2P struct {
 	peerstore pstore.Peerstore
 }
 
+// NewP2P creates new P2P struct
+func NewP2P(identity peer.ID, peerHost p2phost.Host, peerstore pstore.Peerstore) *P2P {
+	return &P2P{
+		identity:  identity,
+		peerHost:  peerHost,
+		peerstore: peerstore,
+
+		Listeners: ListenerRegistry{
+			Listeners: map[string]Listener{},
+		},
+		Streams: StreamRegistry{
+			Streams: map[uint64]*Stream{},
+		},
+	}
+}
+
 // CheckProtoExists checks whether a proto handler is registered to
 // mux handler
 func (p2p *P2P) CheckProtoExists(proto string) bool {
